Add files via upload
To model the price of cars with the available independent variables. It will be used by the automobile company Geely Auto to understand how exactly the prices vary with the independent variables. They can accordingly manipulate the design of the cars, the business strategy etc. to meet certain price levels.